+++ gear/form_qc_checker/src/python/form_qc_app/run.py
@@ -6,6 +6,7 @@ from typing import Optional
 
 from configs.ingest_configs import load_form_ingest_configurations
 from flywheel_gear_toolkit import GearToolkitContext
+from fw_gear import GearContext
 from gear_execution.gear_execution import (
     ClientWrapper,
     GearBotClient,
@@ -60,7 +61,7 @@ class FormQCCheckerVisitor(GearExecutionEnvironment):
 
     @classmethod
     def create(
-        cls, context: GearToolkitContext, parameter_store: Optional[ParameterStore]
+        cls, context: GearContext, parameter_store: Optional[ParameterStore]
     ) -> "FormQCCheckerVisitor":
         """Creates a form-qc-checker execution visitor.
 
@@ -121,7 +122,7 @@ class FormQCCheckerVisitor(GearExecutionEnvironment):
             supplement_input=supplement_input,
         )
 
-    def run(self, context: GearToolkitContext):
+    def run(self, context: GearContext):
         """Runs the form-qc-checker app.
 
         Args:
@@ -131,7 +132,7 @@ class FormQCCheckerVisitor(GearExecutionEnvironment):
         assert context, "Gear context required"
 
         admin_group = self.admin_group(
-            admin_id=context.config.get("admin_group", DefaultValues.NACC_GROUP_ID)
+            admin_id=context.config.opts.get("admin_group", DefaultValues.NACC_GROUP_ID)
         )
 
         try:
@@ -145,6 +146,7 @@ class FormQCCheckerVisitor(GearExecutionEnvironment):
             ) from error
 
         run(
+            gear_name=self.gear_name(context, "form-qc-checker"),
             client_wrapper=self.client,
             input_wrapper=self.__file_input,
             s3_client=self.__s3_client,