+++ gear/form_qc_coordinator/src/python/form_qc_coordinator_app/pipelines.py
@@ -6,7 +6,7 @@ from configs.ingest_configs import FormProjectConfigs, ModuleConfigs, PipelineTy
 from flywheel.models.file_entry import FileEntry
 from flywheel_adaptor.flywheel_proxy import FlywheelProxy
 from flywheel_adaptor.subject_adaptor import ParticipantVisits, SubjectAdaptor
-from flywheel_gear_toolkit import GearToolkitContext
+from fw_gear import GearContext
 from gear_execution.gear_execution import GearExecutionError
 from gear_execution.gear_trigger import GearInfo
 from keys.keys import DefaultValues, MetadataKeys
@@ -24,7 +24,7 @@ class PipelineProcessor(ABC):
         self,
         *,
         proxy: FlywheelProxy,
-        gear_context: GearToolkitContext,
+        gear_context: GearContext,
         subject: SubjectAdaptor,
         module: str,
         visits_info: ParticipantVisits,