+++ gear/form_scheduler/src/python/form_scheduler_app/run.py
@@ -8,7 +8,7 @@ from configs.ingest_configs import ConfigsError, PipelineConfigs
 from event_capture.event_capture import VisitEventCapture
 from flywheel.rest import ApiException
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
-from flywheel_gear_toolkit import GearToolkitContext
+from fw_gear import GearContext
 from gear_execution.gear_execution import (
     ClientWrapper,
     GearBotClient,
@@ -60,7 +60,7 @@ class FormSchedulerVisitor(GearExecutionEnvironment):
     @classmethod
     def create(
         cls,
-        context: GearToolkitContext,
+        context: GearContext,
         parameter_store: Optional[ParameterStore] = None,
     ) -> "FormSchedulerVisitor":
         """Creates a gear execution object.
@@ -85,12 +85,13 @@ class FormSchedulerVisitor(GearExecutionEnvironment):
         )
         assert form_configs_input, "missing expected input, form_configs_file"
 
-        source_email = context.config.get("source_email", "nacchelp@uw.edu")
+        options = context.config.opts
+        source_email = options.get("source_email", "nacchelp@uw.edu")
 
         portal_url = None
         if source_email:
             try:
-                portal_path = context.config.get("portal_url_path", None)
+                portal_path = options.get("portal_url_path", None)
                 if not portal_path:
                     raise GearExecutionError(
                         "No portal URL found, required " + "to send emails"
@@ -99,8 +100,8 @@ class FormSchedulerVisitor(GearExecutionEnvironment):
             except ParameterError as error:
                 raise GearExecutionError(f"Parameter error: {error}") from error
 
-        event_bucket_name = context.config.get("event_bucket", "submission-events")
-        event_environment = context.config.get("event_environment", "prod")
+        event_bucket_name = options.get("event_bucket", "submission-events")
+        event_environment = options.get("event_environment", "prod")
 
         try:
             event_bucket = S3BucketInterface.create_from_environment(event_bucket_name)
@@ -120,10 +121,10 @@ class FormSchedulerVisitor(GearExecutionEnvironment):
             portal_url=portal_url,
         )
 
-    def run(self, context: GearToolkitContext) -> None:
+    def run(self, context: GearContext) -> None:
         """Runs the Form Scheduler app."""
         try:
-            dest_container: Any = context.get_destination_container()
+            dest_container: Any = context.config.get_destination_container()
         except ApiException as error:
             raise GearExecutionError(
                 f"Cannot find destination container: {error}"
@@ -139,8 +140,9 @@ class FormSchedulerVisitor(GearExecutionEnvironment):
 
         # check for other form-scheduler gear jobs running on this project
         # there shouldn't be any
-        gear_name = context.manifest.get("name", "form-scheduler")
-        job_id = context.config_json.get("job", {}).get("id")
+        gear_name = self.gear_name(context, "form-scheduler")
+        job_id = context.config.job.get("id")
+
         if JobPoll.is_another_gear_instance_running(
             proxy=self.proxy,
             gear_name=gear_name,