+++ gear/push_template/src/python/template_app/run.py
@@ -3,8 +3,8 @@
 import logging
 from typing import Optional
 
-from flywheel_gear_toolkit import GearToolkitContext
 from fw_client.client import FWClient
+from fw_gear import GearContext
 from gear_execution.gear_execution import (
     ClientWrapper,
     ContextClient,
@@ -42,7 +42,7 @@ class TemplatingVisitor(GearExecutionEnvironment):
     @classmethod
     def create(
         cls,
-        context: GearToolkitContext,
+        context: GearContext,
         parameter_store: Optional[ParameterStore] = None,
     ) -> "TemplatingVisitor":
         """Creates a templating execution visitor.
@@ -60,22 +60,25 @@ class TemplatingVisitor(GearExecutionEnvironment):
         # ViewerApp type used for copying viewer apps from template projects.
         api_key = get_api_key(context)
         client.set_fw_client(FWClient(api_key=api_key, client_name="push-template"))
-        group_id = context.config.get("template_group")
+
+        options = context.config.opts
+
+        group_id = options.get("template_group")
         if not group_id:
             raise GearExecutionError('Expected "template_group"')
-        template_label = context.config.get("template_project")
+        template_label = options.get("template_project")
         if not template_label:
             raise GearExecutionError('Expected "template_project"')
 
         return TemplatingVisitor(
-            admin_id=context.config.get("admin_group", "nacc"),
+            admin_id=options.get("admin_group", "nacc"),
             client=client,
             template_group=group_id,
             template_label=template_label,
-            new_only=context.config.get("new_only", False),
+            new_only=options.get("new_only", False),
         )
 
-    def run(self, context: GearToolkitContext) -> None:
+    def run(self, context: GearContext) -> None:
         projects = self.proxy.find_projects(
             group_id=self.__template_group, project_label=self.__template_label
         )